pyproject.toml
[project]
name = "promptory-server"
version = "0.1.0"
description = "Git-native LLM prompt management platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "pygithub>=2.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "bcrypt>=4.2.0",
    "cryptography>=43.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
